requirements.txt
you-get
azure-cognitiveservices-speech
python-dotenv
requests 